fix: possibly fix crashes with the two-finger long-press gesture,
it seems that the workaround that explicitely set the gesture sequence state
which was at the time needed to avoid segfaults caused by Gtk
is no longer necessary and the issue has been resolved.

Because this workaround itself apparently caused crashes on certain systems
(mainly on hyprland?), removing the entire code section should fix it.