fix error and API change
Note : the task HAS to exist until the pen is changed to another tool.

This is the condition for stopping to check the condition.

This means that for each long hold, a test is made to see if we can recognize a shape (or more).

If not, then we should reset both methods : start again from scratch on the long hold by event and same thing for the task handle